ARM: dts: imx7d: cl-som-imx7: add support for NAND flash
Add support for NAND flash on GPMI interface.
The NAND option is mutually exlusive with eMMC option,
connected to USBHC3 (SD3) interface. This because GPMI
and USDHC3 share same pins.
The GPMI interface is disabled by default.
It supposed to be enabled via fdt command line utility
in the u-boot.

Signed-off-by: Ilya Ledvich <ilya@compulab.co.il>

%% original patch: 0013-ARM-dts-imx7d-cl-som-imx7-add-support-for-NAND-flash.patch