ARM: dts: imx7d: sbc-iot-imx7: add basic board support
The SB-IOT-iMX7 base board together with CL-SOM-iMX7 SoM form
SBC-IOT-iMX7 single board computer.
SBC-IOT-iMX7 is a single board computer optimized for industrial
control and monitoring, extensive wireless and wired connectivity,
ideal solution for cost-sensitive systems. It is based on the
Freescale i.MX7 system-on-chip. SBC-IOT-iMX7 is implemented with
the CL-SOM-iMX7 System-on-Module providing most of the functions,
and SB-IOT-iMX7 carrier board providing additional peripheral
functions and connectors.

http://www.compulab.com/products/sbcs/sbc-iot-imx7-nxp-i-mx-7-internet-of-things-single-board-computer/

http://www.compulab.co.il/products/computer-on-modules/cl-som-imx7-freescale-i-mx-7-system-on-module/

Add basic board support, including SD card as a secondary boot and
storage device.

Signed-off-by: Ilya Ledvich <ilya@compulab.co.il>

%% original patch: 0032-ARM-dts-imx7d-sbc-iot-imx7-add-basic-board-support.patch