iio:max11108: Fix name of trigger to be allocated
On probe method the ADC driver allocates a trigger. Allocation of the
trigger with a constant name restricts the system to use  only one ADC.
In order to allow usage of more than one ADC, the trigger name has been
suffixed with an unique ID, per sample of other drivers.

Signed-off-by: Kirill Kapranov <kirill.kapranov@compulab.co.il>